UNSTABLE COMMIT: Attempting to patch GPU errors
This commit will be unable to build, and is being pushed for the
purpose of getting others to help debug. The previous commit is
"stable" in that it will build, but fails to fit when using GPUs
as ComplexCoeff is not setup properly. Changes made in this commit
appear to alleviate those issues, but now break something in the
GPU manager when attempting to build for both mpi or gpu. Oddly
enough building for mpigpu works