Initial Commit of simple vecps moment method
The intent of this branch is to add a feature for
the simple implementation to calculate the
vector-pseudoscalar moments from data. "Simple"
here refers to the fact that we will simply be
multiplying the angular distributions by a free
parameter (the moment), and letting MINUIT adjust
these parameters to fit the data. It will likely
be very slow, but the intention is not for long
term use, but rather a stepping stone towards more
proper methods like Boris' work in the
two pseudoscalar case.

The file being uploaded is originally authored by
Edmundo Barriga (edbarriga).