Changed moment index to a struct
Moment index method seemed lofty and hard to
interpret, so it was changed to a struct that
contains all the moment-related parameters.
Aside from variable and comment adjustments, a
bug was also fixed where the H2 moment did not
have the sine of the angle, but mistakenly the
cosine, thus making it real instead of imaginary.