Removed bracket handling of args, and fixed leak
A memory leak was being caused due to the
addresses of struct members being registered, as
opposed to the structs stored in the vector. This
was only 1/2 of the problem, but the other half
seems to lie within AmpTools itself.

I've also changed the args to not remove the
brackets "[]" of parameters, as there is no need
and it prevents any confusion with how the
parameters are handled