Added flag for moment "amplitudes"
The vec_ps moments currently only work for the
omega -> 3pi decay, but the normal flag to detect
this only works if "omega3pi" is in the cfg file,
which won't be the case for these simple moment
fits. This commit ensures that the flag is turned
on if moments are being used.

Ideally a separate plotter is made for the moments
to see their angular distribution, but for now
this is a quick way to check that the phasespace
is being weighted properly.